Feat: BaseCollectionViewCell 생성/적용
